Struct strck::Check

#[repr(transparent)]
pub struct Check<I: Invariant, B: AsRef<str> + 'static = String> {
    pub(crate) _marker: PhantomData<I>,
    pub(crate) buf: B,
}

Owned immutable string with invariants.

Similar to how String derefs to &str, Check derefs to &Ck. This means APIs requiring &Check<I> as an argument should instead consider accepting &Ck<I> for more flexibility.

Buffers

By default, this type is backed by a String, but it can also be backed by any AsRef<str> + 'static type. In particular, types like SmolStr are good candidates since they’re designed to be immutable.

It’s recommended to use a type alias when using a custom backing type, since extra generics can make the type signature long.

Fields
_marker: PhantomData<I>buf: B
Implementations

impl<I: Invariant, B: AsRef<str>> Check<I, B>

pub fn from_buf(buf: B) -> Result<Self, I::Error>

Returns an Ok if the buffer upholds the invariants, otherwise Err.

pub unsafe fn from_buf_unchecked(buf: B) -> Self

Create a new Check without validating the buffer.

Safety

The buffer must contain a valid string.

pub fn as_ck(&self) -> &Ck<I>

Returns a &Ck that borrows from self.

pub fn into_inner(self) -> B

Returns the inner representation.